admitad_test/pdf_generator/serializers/generic_serializer.py
from rest_framework.serializers import Serializer


class GenericSourceToHtmlSerializer(Serializer):
    def get_fields(self):
        raise NotImplementedError

    def create(self, validated_data):
        raise NotImplementedError
